ASoC: SOF: Modify the signature of load_firmware op
Modify the signature of the load_firmware op to take the firmware
filename as an argument. The firmware filename should be include the
full path of the firmware file to be loaded. This is in preparation
to add a debug op that will be used to boot a firmware with the
filename passed from userspace.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>