ASoC: SOF: Intel: hda: Move DSP power step to the set_power_state op
This will allow for the set_power_state op to be used for switching the
DSP to the D3 state. The logic for powering up the DSP when setting the
state to D0 also needs to be moved and will be done in a follow up. But
for now, this will allow stress testing firmware boot.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>